backend/pyproject.toml
[project]
name = "isw-backend"
version = "0.1.0"
description = "ISW Backend Project"
requires-python = ">=3.11"
dependencies = [
    "django==4.2.20",
    "djangorestframework==3.15.2",
    "drf-nested-routers==0.94.1",
    "redis==3.5.3",
    "django-redis==5.1.0",
    "django-environ==0.4.5",
    "django-extensions==3.1.5",
    "django-cors-headers==3.10.1",
    "cryptography==40.0.2",
    "argon2-cffi==23.1.0",
    "celery==5.2.7",
    "openpyxl==3.0.10",
    "django-filter==21.1",
    "djangorestframework-simplejwt==5.3.1",
    "sphinx==2.0.1",
    "psycopg2-binary==2.9.9",
    "django-coverage-plugin==1.6.0",
    "factory-boy==2.12.0",
    "cos-python-sdk-v5==1.9.36",
    "qcloud-python-sts==3.0.5",
    "xlrd==1.2.0",
    "pycryptodome==3.20.0",
    "django-mptt==0.14.0",
    "pyyaml==6.0.1",
    "django-requestlogs==0.7.1",
    "drf-dynamic-fields==0.4.0",
    "daphne==4.0.0",
    "paho-mqtt==1.5.1",
    "django-celery-beat==2.5.0",  # 如果安装失败，请使用: uv pip install --no-binary django-celery-beat django-celery-beat==2.5.0
    "drf-spectacular==0.27.0",
    "chinesecalendar==1.10.0",
    "requests==2.32.3",
    "django-utils-six==2.0",
    "influxdb-client==1.40.0",
    "PyExecJS==1.5.1",
    "tencentcloud-sdk-python==3.0.1015",
    "werkzeug==2.2.3",
    "twisted==22.10.0",
    "pillow==10.1.0",
    "psycopg2==2.9.7",
    "django-postgres-extra==2.0.8",
    "watchdog==3.0.0",
    "func-timeout==4.3.5",
    "js2py==0.74",
    "py-mini-racer==0.6.0",
    "taospy==2.7.21",
    "taos-ws-py==0.3.8",
    "numpy==1.24.4",
    "concurrent-log-handler==0.9.26",
    "crcmod==1.7",
    "dashscope==1.24.2",
    "uvicorn==0.35.0",
    "adrf==0.1.9",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar==3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

